Initial multi-branch support
This change makes multi-branch support possible in Jenkins.  It extends the list of of configuration properties for a repo to a branch name.  If a branch name is specified, a branch specific folder is generated underneath the repo name.  GithubBranchName is passed to netci.groovy while generating, and should be used when setting up jobs (Utilities.standardJobSetup, addStandardParameters)