Customize defaults for GitBranchOrCommit and GitRefSpec
This change allows callers to customize the default values for GitBranchOrCommit and GitRefSpec.

The previous attempt at this change modified the signatures of existing methods in place and subtly changed the defaults.  This approach defines new methods and has the existing ones feed into those.  It also use a more declarative model that should be more resilient to changes.

Verified that locally generated PR jobs still have '${sha1}' as the push value.