Allow for multiple DSL definitions
This change allows for multiple groovy files to be used. They can also target different servers if desired.  This is motivated by some teams' desires to move CI definitions into subdirectories, as well as a desire to have a separate perf server.  We can put the perf definitions in another file which runs on the other server.